backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
httpx==0.25.2
tenacity==8.2.3
cryptography==41.0.8
bcrypt==4.1.2
pyjwt==2.8.0
psutil==5.9.6 